Move node related pkg/util to component-helpers
