+++ pkg/proxy/util/utils.go
@@ -31,7 +31,7 @@ import (
 	"k8s.io/apimachinery/pkg/util/sets"
 	"k8s.io/client-go/tools/record"
 	helper "k8s.io/component-helpers/networking/corev1"
-	utilsysctl "k8s.io/kubernetes/pkg/util/sysctl"
+	utilsysctl "k8s.io/component-helpers/node/utils/sysctl"
 	utilnet "k8s.io/utils/net"
 
 	"k8s.io/klog/v2"
