+++ pkg/proxy/winuserspace/proxier.go
@@ -32,9 +32,9 @@ import (
 	utilnet "k8s.io/apimachinery/pkg/util/net"
 	"k8s.io/apimachinery/pkg/util/runtime"
 	helper "k8s.io/component-helpers/networking/corev1"
+	"k8s.io/component-helpers/networking/utils/netsh"
 	"k8s.io/kubernetes/pkg/proxy"
 	"k8s.io/kubernetes/pkg/proxy/config"
-	"k8s.io/kubernetes/pkg/util/netsh"
 )
 
 const allAvailableInterfaces string = ""
