+++ pkg/proxy/winuserspace/proxysocket.go
@@ -30,9 +30,9 @@ import (
 	v1 "k8s.io/api/core/v1"
 	"k8s.io/apimachinery/pkg/types"
 	"k8s.io/apimachinery/pkg/util/runtime"
+	"k8s.io/component-helpers/networking/utils/ipconfig"
 	"k8s.io/klog/v2"
 	"k8s.io/kubernetes/pkg/proxy"
-	"k8s.io/kubernetes/pkg/util/ipconfig"
 	"k8s.io/utils/exec"
 )
